create database core

+++ app/src/main/java/net/gas/contactbook/ui/database/entities/Departments.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "departments")
 data class Departments(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Short,
     val name: String?,
     val code: String?,
     val rangir: Int?
