+++ app/src/main/java/net/gas/contactbook/ui/database/entities/Units.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "units")
 data class Units (
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Short,
     val name: String?,
     val code: String?,
     val rangir: Int?
